refactor: 연관있는 데이터를 기준으로 커스텀 훅 리팩토링(선택한 주문 id와 상태 저장, 주문 상태 변경 mutation 으로 분리)
